Move routers/api/v1/utils/utils PaginateSlice -> modules/util/paginate.go
